requirements.txt
flask
flask-socketio
pycryptodome